requirements.txt
Django==2.0.7
djangorestframework==3.8.2
pytz==2018.5
uWSGI==2.0.17.1
